render.js deleted, new presenter for points
