+++ src/main.js
@@ -1,6 +1,6 @@
 import FiltersView from './view/filters-view.js';
 import EventsPresenter from './presenter/events-presenter.js';
-import { render } from './render.js';
+import { render } from './framework/render.js';
 import PointsModel from './model/points-model.js';
 
 const siteHeaderElement = document.querySelector('.page-header');
