+++ src/model/points-model.js
@@ -3,7 +3,7 @@ import { generatePoint } from '../mocks/point-mock.js';
 
 export default class PointsModel {
 
-  points = Array.from({length: 0}, generatePoint);
+  points = Array.from({length: 10}, generatePoint);
   getPoints = () => this.points;
 
   getDestinations = () => destinations;
