+++ src/presenter/events-presenter.js
@@ -3,9 +3,10 @@ import SortView from '../view/sort-view.js';
 import EventEditorView from '../view/event-editor-view.js';
 import PointInListView from '../view/point-in-list.js';
 import EventsUlistView from '../view/events-ulist-view.js';
-import { render } from '../render.js';
+import { render } from '../framework/render.js';
 import EmptyListView from '../view/empty-list-view.js';
 import { filterTypes } from '../view/filters-view.js';
+import PointPresenter from './point-presenter.js';
 
 export default class EventsPresenter {
 
@@ -32,41 +33,9 @@ export default class EventsPresenter {
         const eventPoint = new PointInListView(this.eventPoints[i],
           this.eventDestinations[i],
           this.eventOffers[i]);
-        this.#renderPoint(eventPoint, editPoint);
+        const pointPresenter = new PointPresenter(editPoint, eventPoint);
+        pointPresenter.renderPoint(this.#eventsUlistView);
       }
     }
-  };
-
-  #renderPoint = (eventPoint, editPoint) => {
-
-    const replacePointToEdit = (point, edit) => {
-      this.#eventsUlistView.element.replaceChild(edit.element, point.element);
-    };
-
-    const replaceEditToPoint = (point, edit) => {
-      this.#eventsUlistView.element.replaceChild(point.element, edit.element);
-    };
-
-    const onEscKeyDown = (evt) => {
-      if (evt.key === 'Escape' || evt.key === 'Esc') {
-        evt.preventDefault();
-        replaceEditToPoint(eventPoint, editPoint);
-        document.removeEventListener('keydown', onEscKeyDown);
-      }
-    };
-
-    eventPoint.setEditClickHandler(() => {
-      replacePointToEdit(eventPoint, editPoint);
-      document.addEventListener('keydown', onEscKeyDown);
-    });
-    editPoint.setRollupHandler(() => {
-      replaceEditToPoint(eventPoint, editPoint);
-      document.removeEventListener('keydown', onEscKeyDown);
-    });
-    editPoint.setFormSubmitHandler(() => {
-      replaceEditToPoint(eventPoint, editPoint);
-      document.removeEventListener('keydown', onEscKeyDown);
-    });
-    render(eventPoint, this.#eventsUlistView.element);
   };
 }
